First Commit!
Created outline for HTML and CSS indexes.
Created file system.
Added Header with logo, 2 hero sections, and footer
Attempted to "freehand" layout in CSS using HW doc as a rough outline that didn't quite work because I tried to follow the layout of my own portfolio rather than the default suggested in the assignment; still needs a lot of work.